feature-BuyCryptocurrencies done!

+++ src/main/java/org/clerodri/entity/CryptoMoney.java
@@ -1,9 +1,7 @@
 package org.clerodri.entity;
 
-import org.clerodri.CryptoEnum;
-
 public class CryptoMoney {
-    private Integer quantity;
+    private double quantity;
     private Integer value;
 
     public CryptoMoney( Integer quantity, Integer value) {
@@ -11,11 +9,11 @@ public class CryptoMoney {
         this.value = value;
     }
 
-    public Integer getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
